Feature/control plane

+++ include/log_entry.h
@@ -7,7 +7,7 @@ struct ChatEntry {
     std::string message_id;
     std::string sender_id;
     std::string payload;
-    int32_t     lamport_time;
+    int64_t     lamport_time;
     int32_t     epoch;
     uint32_t    checksum;
 };
